+++ src/view/canvas_components/canvas.ts
@@ -133,10 +133,10 @@ export class ViewCanvas {
     }
   }
 
-  private AddGrid(byFabric = false) {
-    // by image
-    if (!byFabric) {
-      const img = GetImage(ImageHandle.GRID);
+  private AddGrid() {
+    const img = GetImage(ImageHandle.GRID);
+    // TODO: 'onload' should be removed after we can pre-load all assets.
+    img.onload = () => {
       const obj = new fabric.Image(img, {
         originX: 'center',
         originY: 'center',
@@ -145,31 +145,7 @@ export class ViewCanvas {
         evented: false,
       });
       this.canvas.add(obj);
-    } else {
-      const options = {
-        distance: 10,
-        width: 4000,
-        height: 4000,
-        param: {stroke: '#E0EBEB', strokeWidth: 1, evented: false}
-      },
-
-            gridLen = options.width / options.distance;
-      for (let i = -gridLen / 2; i < gridLen / 2; i++) {
-        const distance = i * options.distance,
-              horizontal = new fabric.Line(
-                  [distance, -options.width / 2, distance, options.width / 2],
-                  options.param),
-              vertical = new fabric.Line(
-                  [-options.width / 2, distance, options.width / 2, distance],
-                  options.param);
-        this.canvas.add(horizontal);
-        this.canvas.add(vertical);
-        if (i % 5 === 0) {
-          horizontal.set({stroke: '#7AA7A7'});
-          vertical.set({stroke: '#7AA7A7'});
-        }
-      }
-    }
+    };
   }
 
   private EnableZoom() {
